requirements.txt
python-dotenv-1.0.1
psycopg2-2.9.10
pyyaml-6.0.2
openai-0.28